requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.5.0=pypi_0
anyio=3.7.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.0.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
bs4=0.0.1=pypi_0
build=0.10.0=pypi_0
ca-certificates=2023.05.30=hecd8cb5_0
cachecontrol=0.13.1=pypi_0
certifi=2023.7.22=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.2.0=pypi_0
chroma-hnswlib=0.7.2=pypi_0
chromadb=0.4.7=pypi_0
cleo=2.0.1=pypi_0
click=8.1.7=pypi_0
coloredlogs=15.0.1=pypi_0
crashtest=0.4.1=pypi_0
dataclasses-json=0.5.14=pypi_0
datasets=2.14.4=pypi_0
dill=0.3.7=pypi_0
distlib=0.3.7=pypi_0
docx2txt=0.8=pypi_0
dulwich=0.21.6=pypi_0
exceptiongroup=1.1.3=pypi_0
fastapi=0.99.1=pypi_0
filelock=3.12.2=pypi_0
flatbuffers=23.5.26=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
google-search-results=2.4.2=pypi_0
graphlib-backport=1.0.3=pypi_0
greenlet=2.0.2=pypi_0
h11=0.14.0=pypi_0
httptools=0.6.0=pypi_0
huggingface-hub=0.16.4=pypi_0
humanfriendly=10.0=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.1=pypi_0
installer=0.7.0=pypi_0
jaraco-classes=3.3.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.17.3=pypi_0
keyring=24.2.0=pypi_0
langchain=0.0.309=pypi_0
langsmith=0.0.43=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_0
lxml=4.9.3=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.20.1=pypi_0
monotonic=1.6=pypi_0
more-itertools=10.1.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.7=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=hcec6c5f_0
networkx=3.1=pypi_0
nltk=3.8.1=pypi_0
numexpr=2.8.5=pypi_0
numpy=1.24.4=pypi_0
onnxruntime=1.15.1=pypi_0
openai=0.27.9=pypi_0
openssl=3.0.10=hca72f7f_1
overrides=7.4.0=pypi_0
pandas=2.0.3=pypi_0
pexpect=4.8.0=pypi_0
pillow=9.5.0=pypi_0
pip=23.2.1=py38hecd8cb5_0
pkginfo=1.9.6=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.11.0=pypi_0
poetry=1.6.1=pypi_0
poetry-core=1.7.0=pypi_0
poetry-plugin-export=1.5.0=pypi_0
posthog=3.0.2=pypi_0
protobuf=3.20.0=pypi_0
ptyprocess=0.7.0=pypi_0
pulsar-client=3.2.0=pypi_0
pyarrow=13.0.0=pypi_0
pycparser=2.21=pypi_0
pydantic=1.10.12=pypi_0
pydantic-core=2.6.3=pypi_0
pypdf2=3.0.1=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.0.0=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.8.17=h5ee71fb_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
python-multipart=0.0.6=pypi_0
python-pptx=0.6.22=pypi_0
pytz=2023.3=pypi_0
pyyaml=6.0.1=pypi_0
ragas=0.0.11=pypi_0
rapidfuzz=2.15.2=pypi_0
readline=8.2=hca72f7f_0
regex=2023.8.8=pypi_0
requests=2.31.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
safetensors=0.3.3=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.10.1=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.0.0=py38hecd8cb5_0
shellingham=1.5.3=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlalchemy=2.0.20=pypi_0
sqlite=3.41.2=h6c40b1e_0
starlette=0.27.0=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
threadpoolctl=3.2.0=pypi_0
tiktoken=0.4.0=pypi_0
tk=8.6.12=h5d9f67b_0
tokenizers=0.13.3=pypi_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.66.1=pypi_0
transformers=4.32.0=pypi_0
trove-classifiers=2023.9.19=pypi_0
typing-extensions=4.7.1=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=2.0.4=pypi_0
uvicorn=0.23.2=pypi_0
uvloop=0.17.0=pypi_0
virtualenv=20.24.5=pypi_0
watchfiles=0.20.0=pypi_0
websockets=11.0.3=pypi_0
wheel=0.38.4=py38hecd8cb5_0
xattr=0.10.1=pypi_0
xlsxwriter=3.1.6=pypi_0
xxhash=3.3.0=pypi_0
xz=5.4.2=h6c40b1e_0
yarl=1.9.2=pypi_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h4dc903c_0
